refactor: update donation form model and repository
