+++ src/DonationForms/V2/DataTransferObjects/DonationFormQueryData.php
@@ -3,10 +3,14 @@
 namespace Give\DonationForms\V2\DataTransferObjects;
 
 use DateTime;
+use Give\DonationForms\Properties\FormSettings;
+use Give\DonationForms\Properties\GoalSettings;
 use Give\DonationForms\V2\Models\DonationForm;
 use Give\DonationForms\V2\Properties\DonationFormLevel;
-use Give\DonationForms\V2\ValueObjects\DonationFormMetaKeys;
+use Give\DonationForms\V2\ValueObjects\DonationFormMetaKeys as LegacyDonationFormMetaKeys;
 use Give\DonationForms\V2\ValueObjects\DonationFormStatus;
+use Give\DonationForms\ValueObjects\DonationFormMetaKeys;
+use Give\DonationForms\ValueObjects\GoalType;
 use Give\Framework\Support\Facades\DateTime\Temporal;
 use Give\Framework\Support\ValueObjects\Money;
 
@@ -56,6 +60,16 @@ final class DonationFormQueryData
     public $status;
 
     /**
+     * @unreleased
+     */
+    public GoalSettings $goalSettings;
+
+    /**
+     * @unreleased
+     */
+    public bool $usesFormBuilder;
+
+    /**
      * Convert data from donation form object to DonationForm Model
      *
      * @since 2.24.0
@@ -66,17 +80,19 @@ final class DonationFormQueryData
      */
     public static function fromObject($object): DonationFormQueryData
     {
-        $self = new static();
+        $self = new DonationFormQueryData();
 
         $self->id = (int)$object->id;
         $self->title = $object->title;
         $self->levels = $self->getDonationFormLevels($object);
-        $self->goalOption = ($object->{DonationFormMetaKeys::GOAL_OPTION()->getKeyAsCamelCase()} === 'enabled');
+        $self->goalOption = ($object->{LegacyDonationFormMetaKeys::GOAL_OPTION()->getKeyAsCamelCase()} === 'enabled');
         $self->createdAt = Temporal::toDateTime($object->createdAt);
         $self->updatedAt = Temporal::toDateTime($object->updatedAt);
-        $self->totalAmountDonated = Money::fromDecimal($object->{DonationFormMetaKeys::FORM_EARNINGS()->getKeyAsCamelCase()}, give_get_currency());
-        $self->totalNumberOfDonations = (int)$object->{DonationFormMetaKeys::FORM_SALES()->getKeyAsCamelCase()};
+        $self->totalAmountDonated = Money::fromDecimal($object->{LegacyDonationFormMetaKeys::FORM_EARNINGS()->getKeyAsCamelCase()}, give_get_currency());
+        $self->totalNumberOfDonations = (int)$object->{LegacyDonationFormMetaKeys::FORM_SALES()->getKeyAsCamelCase()};
         $self->status = new DonationFormStatus($object->status);
+        $self->goalSettings = $self->getGoalSettings($object);
+        $self->usesFormBuilder = (bool)$object->settings;
 
         return $self;
     }
@@ -102,9 +118,9 @@ final class DonationFormQueryData
      */
     public function getDonationFormLevels($object): array
     {
-        switch( $object->{DonationFormMetaKeys::PRICE_OPTION()->getKeyAsCamelCase()} ) {
+        switch( $object->{LegacyDonationFormMetaKeys::PRICE_OPTION()->getKeyAsCamelCase()} ) {
             case 'multi':
-                $levels = maybe_unserialize($object->{DonationFormMetaKeys::DONATION_LEVELS()->getKeyAsCamelCase()});
+                $levels = maybe_unserialize($object->{LegacyDonationFormMetaKeys::DONATION_LEVELS()->getKeyAsCamelCase()});
 
                 if (empty($levels)) {
                     return [];
@@ -114,7 +130,7 @@ final class DonationFormQueryData
                     return DonationFormLevel::fromArray($level);
                 }, $levels);
             case 'set':
-                $amount = $object->{DonationFormMetaKeys::SET_PRICE()->getKeyAsCamelCase()};
+                $amount = $object->{LegacyDonationFormMetaKeys::SET_PRICE()->getKeyAsCamelCase()};
 
                 if (empty($amount)) {
                     return [];
@@ -125,6 +141,60 @@ final class DonationFormQueryData
                 ];
             default:
                 return [];
+        }
+    }
+
+
+    /**
+     * @unreleased
+     */
+    private function getGoalSettings(object $queryObject): GoalSettings
+    {
+        $currency = give_get_option('currency', 'USD');
+        $settings = $queryObject->{DonationFormMetaKeys::SETTINGS()->getKeyAsCamelCase()};
+
+        if ($settings) {
+            $settings = FormSettings::fromjson($settings);
+
+            return GoalSettings::fromArray([
+                'enableDonationGoal' => $settings->enableDonationGoal,
+                'goalType' => $settings->goalType,
+                'goalAmount' => $settings->goalType->isOneOf(GoalType::AMOUNT(), GoalType::AMOUNT_FROM_SUBSCRIPTIONS())
+                    ? Money::fromDecimal($queryObject->goalAmount, $currency)->formatToDecimal()
+                    : $settings->goalAmount,
+            ]);
+        }
+
+        $goalType = $this->convertGoalType($queryObject->goalFormat, (bool)$queryObject->recurringGoalFormat);
+
+        return GoalSettings::fromArray([
+            'enableDonationGoal' => $queryObject->goalOption === 'enabled',
+            'goalType' => $goalType,
+            'goalAmount' => $goalType->isOneOf(GoalType::AMOUNT(), GoalType::AMOUNT_FROM_SUBSCRIPTIONS())
+                ? Money::fromDecimal($queryObject->goalAmount, $currency)->formatToDecimal()
+                : $settings->goalAmount,
+        ]);
+    }
+
+
+    /**
+     * @unreleased
+     */
+    public function convertGoalType(string $type, bool $isRecurring): GoalType
+    {
+        switch ($type) {
+            case 'donation':
+                return $isRecurring
+                    ? GoalType::SUBSCRIPTIONS()
+                    : GoalType::DONATIONS();
+            case 'donors':
+                return $isRecurring
+                    ? GoalType::DONORS_FROM_SUBSCRIPTIONS()
+                    : GoalType::DONORS();
+            default:
+                return $isRecurring
+                    ? GoalType::AMOUNT_FROM_SUBSCRIPTIONS()
+                    : GoalType::AMOUNT();
         }
     }
 }
