+++ src/DonationForms/V2/Models/DonationForm.php
@@ -4,7 +4,7 @@ namespace Give\DonationForms\V2\Models;
 
 use DateTime;
 use Give\DonationForms\Models\DonationForm as ModelsDonationForm;
-use Give\DonationForms\Properties\FormSettings;
+use Give\DonationForms\Properties\GoalSettings;
 use Give\DonationForms\V2\Actions\ConvertV2FormToV3Form;
 use Give\DonationForms\V2\DataTransferObjects\DonationFormQueryData;
 use Give\DonationForms\V2\Properties\DonationFormLevel;
@@ -29,6 +29,10 @@ use Give\Framework\Support\ValueObjects\Money;
  * @property DateTime $createdAt
  * @property DateTime $updatedAt
  * @property DonationFormStatus $status
+ * @unreleased
+ * @property GoalSettings $goalSettings
+ * @property bool $enableDonationGoal
+ * @property bool $usesFormBuilder
  */
 class DonationForm extends Model implements ModelReadOnly
 {
@@ -45,6 +49,9 @@ class DonationForm extends Model implements ModelReadOnly
         'createdAt' => DateTime::class,
         'updatedAt' => DateTime::class,
         'status' => DonationFormStatus::class,
+        'goalSettings' => GoalSettings::class,
+        'enableDonationGoal' => 'bool',
+        'usesFormBuilder' => 'bool',
     ];
 
     /**
